apps/sandbox-worker/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
pandas==2.1.4
numpy==1.26.3
yfinance>=0.2.40
scipy==1.11.4
statsmodels==0.14.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.1
